requirements.txt
Flask
pandas
numpy
scikit-learn
Flask-Cors
matplotlib
seaborn
statsmodels
openpyxl
pymongo
python-dotenv
pyyaml
dill
-e .